ReHM_dashboard/requirements.txt
Django==4.0.5
django-bootstrap-v5==1.0.11
django-filter==21.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
whitenoise==6.1.0
django-filter==21.1
channels==3.0.4
channels-redis==3.4.0
black==22.3.0
djongo==1.3.6
pymongo[srv]
django-livereload-server
fastapi==0.78.0
uvicorn[standard]
paho-mqtt==1.6.1
asyncio-mqtt